Fixes to the wchar conversions for radioname and dmrid displays. Thanks to @iddq for ripping his eyeballs out and pointing me in the right direction.
